requirements.txt
Flask
peewee
requests
Flask-Redis
Werkzeug
psycopg2-binary
rq
worker